% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_SUMCA.R
\name{mspe_FH_Sumca}
\alias{mspe_FH_Sumca}
\title{MSPE estimation in FH model using SUMCA method. Calculate the mspe of Fay-Herriot model in SAE using Sumca method.}
\usage{
mspe_FH_Sumca(m, p, X, beta, A, D, K, R)
}
\arguments{
\item{m}{number of small areas}

\item{p}{number of fixed model parameters}

\item{X}{covariates}

\item{beta}{regression coefficients}

\item{A}{variance of area-specific random effects}

\item{D}{sampling variance}

\item{K}{number of Monte Carlo for the SUMCA method}

\item{R}{number of simulation runs}
}
\value{
Par: return estimation of model parameters

MSPE.TRUE.Final: return empirical MSPE of small area predictor

mspe.Sumca.Final: return mspe of small area predictor using the SUMCA method

RB.SUMCA: return relative bias (RB) of mspe of small area predictor using the SUMCA method
}
\description{
MSPE estimation in FH model using SUMCA method. Calculate the mspe of Fay-Herriot model in SAE using Sumca method.
}
\examples{
mspe_FH_Sumca(20,3,matrix(runif(60,0,1),nrow=20,byrow=TRUE),c(1,1,1),10,2.5,10,10)


}
