% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthData_Logistic_JLW.R
\name{mspe_LOGISTIC_HealthData_JLW}
\alias{mspe_LOGISTIC_HealthData_JLW}
\title{MSPE estimation in mixed logistic model (Health Insurance data) using jackknife method.Calculate the mspe of mixed logistic model (Health Insurance data) using jackknife method.}
\usage{
mspe_LOGISTIC_HealthData_JLW(
  m,
  p,
  n.new,
  y.new,
  Xi,
  yi.tem,
  cum.n.new,
  county.tem,
  X.tem
)
}
\arguments{
\item{m}{number of domains}

\item{p}{number of complete model parameters}

\item{n.new}{sample size of each domain}

\item{y.new}{response variable}

\item{Xi}{covariates for each domain}

\item{yi.tem}{response variable for each individual}

\item{cum.n.new}{Cummulative sum of n}

\item{county.tem}{county}

\item{X.tem}{Individual level covariates}
}
\value{
Par: return estimation of model parameters

Mu.hat: return prediction of domain parameters

mspe.JLW: return mspe of small area (domain) predictor using the jackknife method

sq.mspe.JLW: return square root of mspe of small area predictor for non-zero domains using the jackknife method
}
\description{
MSPE estimation in mixed logistic model (Health Insurance data) using jackknife method.Calculate the mspe of mixed logistic model (Health Insurance data) using jackknife method.
}
\examples{
mspe_LOGISTIC_HealthData_JLW(20,3,c(2,1,2,2,1,2,3,1,1,3,1,3,2,3,3,
1,2,1,3,3),c(3,4,2,2,3,3,4,3,4,1,4,1,3,5,4,7,1,3,1,2),
matrix(runif(60,0,1),nrow=20,byrow=TRUE),sample(c(0,1),replace=TRUE,40),
c(2,3,5,7,8,10,13,14,15,18,19,22,24,27,30,31,33,34,37,40),rep(1:20,each=2),
matrix(c(runif(40,7,10),runif(40,14,22),runif(40,2,4)),nrow=40,byrow=FALSE))


}
