% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_Logistic_SUMCA.R
\name{mspe_MS_LOGISTIC_SUMCA}
\alias{mspe_MS_LOGISTIC_SUMCA}
\title{Model selection MSPE estimation in mixed logistic model using SUMCA method. Calculate the model selection mspe of mixed logistic model using SUMCA method.}
\usage{
mspe_MS_LOGISTIC_SUMCA(m, p, ni, X, beta, A, K, R)
}
\arguments{
\item{m}{number of small areas}

\item{p}{number of complete model parameters}

\item{ni}{sample size of each small area}

\item{X}{covariates for the complete model}

\item{beta}{regression coefficients of the complete model}

\item{A}{variance of area-specific random effects}

\item{K}{number of Monte Carlo for the SUMCA method}

\item{R}{number of simulation runs}
}
\value{
Par1: return estimation of model parameters of the complete model

Par2: return estimation of model parameters of the reduced model

MSPE: return empirical MSPE of small area predictor

mspe.Sumca: return mspe of small area predictor using the SUMCA method

RB.SUMCA: return relative bias (RB) of mspe of small area predictor using the SUMCA method

BIC: return BIC of the complete and reduced models
}
\description{
Model selection MSPE estimation in mixed logistic model using SUMCA method. Calculate the model selection mspe of mixed logistic model using SUMCA method.
}
\examples{
mspe_MS_LOGISTIC_SUMCA(20,3,2,matrix(runif(60,0,1),nrow=20,byrow=TRUE),c(1,1,1),10,5,5)

}
